% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlas.R
\name{bufferToRaster}
\alias{bufferToRaster}
\title{Convert buffer to raster}
\usage{
bufferToRaster(buffer)
}
\arguments{
\item{buffer}{The buffer component of a glyph atlas.}
}
\description{
This function converts the buffer component from
a \code{\link{glyphAtlas}} result to a raster object
that can be plotted in R.
}
