% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_text_to_raster.R
\name{draw_text_to_raster}
\alias{draw_text_to_raster}
\title{Draw some text to a raster}
\usage{
draw_text_to_raster(
  text,
  family = par3d("family"),
  color = material3d("color")[1],
  font = par3d("font"),
  cex = par3d("cex")
)
}
\arguments{
\item{text}{A character string to measure}

\item{family, font, cex}{Characteristics of the text}

\item{color}{Foreground color to draw}
}
\value{
A list with components:
* raster, a native raster object.
* locations, a matrix giving the locations of each
  text string in the raster.
}
\description{
Draw some text to a raster
}
\examples{
draw_text_to_raster(c("a", "abc", "j"), family = "serif")
}
