% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawToBitmap.R
\name{drawToBitmap}
\alias{drawToBitmap}
\title{Draw text to bitmap}
\usage{
drawToBitmap(
  texts,
  cex = par3d("cex"),
  family = par3d("family"),
  color = material3d("color")[1],
  font = par3d("font"),
  background = "transparent",
  powerOfTwo = TRUE,
  verbose = FALSE,
  showBaselines = FALSE,
  onePerLine = FALSE,
  minWidth = 0,
  minHeight = 0
)
}
\arguments{
\item{texts}{Text to draw.}

\item{cex, }{family, font Text characteristics.}

\item{background}{Background colour.}

\item{powerOfTwo}{Round bitmap size up to power of two dimensions.}

\item{verbose}{Print progress.}

\item{showBaselines}{Show text baselines (for debugging).}

\item{onePerLine}{Put each bitmap on a new line.}

\item{minWidth, }{minHeight Minimum bitmap measurements.}
}
\value{
A raster object containing the output.
}
\description{
Draw text to bitmap
}
\examples{
text <- c("abc", "Tokyo (\u6771\u4eac)", "\u6771\u4eac", "abc")
bm <- drawToBitmap(text, cex = 3, family = "sans", font = 1,
                  powerOfTwo = FALSE,
                  showBaselines = TRUE, onePerLine = TRUE)
plot.new()
rasterImage(bm, 0, 0, 1, 1)
dim(bm)
attr(bm, "metrics")

}
