% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_text.R
\name{measure_text}
\alias{measure_text}
\title{Measure some text (in pixels)}
\usage{
measure_text(text, family = "sans", font = 1, fontfile = NULL, cex = 1)
}
\arguments{
\item{text}{A character string to measure}

\item{family, font, cex}{Characteristics of the text}

\item{fontfile}{Freetype font file name, or NULL}
}
\value{
A matrix of measurements, one row per entry.
Columns are
\describe{
\item{height, width}{Size of the bounding box for the string}
\item{x_bearing, y_bearing}{The displacement from the starting point to the upper left corner of the bounding box.}
\item{x_advance, y_advance}{Suggested reference point for
a following string.}
\item{ascent, descent}{Amount the fonts used for each
string rise above or descend below the baseline.}
\item{baseline}{How far from the baseline to the top of the bounding box. }
}
}
\description{
Measure some text (in pixels)
}
\examples{
famnum <- rep(1:3, 8)
family <- c("serif", "sans", "mono")[famnum]
font <- rep(rep(1:4, each = 3), 2)
cex <- rep(1:2, each = 12)
measure_text(family, family, font, NULL, cex)
}
