% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack_text.R
\name{pack_text}
\alias{pack_text}
\title{Pack text}
\usage{
pack_text(key, measure, width)
}
\arguments{
\item{key}{Unique key for bitmap, used to detect dups}

\item{measure}{Result of `measure_text()`}

\item{width}{Width in which to pack items}
}
\value{
A matrix of xy values, one per entry
}
\description{
Pack text
}
\examples{
famnum <- rep(1:3, 8)
family <- c("serif", "sans", "mono")[famnum]
font <- rep(rep(1:4, each = 3), 2)
cex <- rep(1:2, each = 12)
m <- measure_text(family, family, font, NULL, cex)
key <- paste(family, font, cex, sep="_")
pack_text(key, m, max(m[, "width"]) + 2)
}
