% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_text_to_raster.R
\name{draw_text_to_raster}
\alias{draw_text_to_raster}
\title{Draw some text to a raster}
\usage{
draw_text_to_raster(text, family, font, fontfile = NULL, cex)
}
\arguments{
\item{text}{A character string to measure}

\item{family, font, cex}{Characteristics of the text}

\item{fontfile}{Freetype font file name, or NULL}
}
\value{
A list with components:
* raster, a native raster object.
* locations, a matrix giving the locations of each
  text string in the raster.
}
\description{
Draw some text to a raster
}
\examples{
famnum <- rep(1:3, 8)
family <- c("serif", "sans", "mono")[famnum]
font <- rep(rep(1:4, each = 3), 2)
cex <- rep(1:2, each = 12)
draw_text_to_raster(family, family, font, NULL, cex) |> image()
}
