% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_text.R
\name{measure_text}
\alias{measure_text}
\title{Measure some text (in pixels)}
\usage{
measure_text(text, family, font, fontfile = NULL, cex)
}
\arguments{
\item{text}{A character string to measure}

\item{family, font, cex}{Characteristics of the text}

\item{fontfile}{Freetype font file name, or NULL}
}
\value{
A matrix of measurements, one row per entry.
Columns are
\describe{
\item{x_bearing, y_bearing}{The displacement from the starting point to the upper left corner of the bounding box.}
\item{width, height}{Size of the bounding box.}
\item{x_advance, y_advance}{Suggested reference point for
a following string.}}
}
\description{
Measure some text (in pixels)
}
\examples{
measure_text(c("a", "abc", "j"), family = "serif",
             font = 1, cex = 1)
}
