% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack_text.R
\name{pack_text}
\alias{pack_text}
\title{Pack text}
\usage{
pack_text(text, measure, width)
}
\arguments{
\item{text}{The text to pack, used to detect dups}

\item{measure}{Result of `measure_text()`}

\item{width}{Width in which to pack items}
}
\value{
A matrix of xy values, one per entry
}
\description{
Pack text
}
\examples{
text <- c("a", "abc", "j")
m <- measure_text(text, family = "serif",
             font = 1, cex = 1)
pack_text(text, m, max(m[, "width"]) + 2)
}
